% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_alpha.R
\name{calibrate_alpha_k}
\alias{calibrate_alpha_k}
\title{Calibration of alpha based on k, the number of clusters}
\usage{
calibrate_alpha_k(fit, alpha)
}
\arguments{
\item{fit}{fitted object using importance sampler. Must have "save_logweight", "save_k", and "cohesion_param$alpha"}

\item{alpha}{vector of alpha values.}
}
\description{
Calibration of alpha based on k, the number of clusters
}
