% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphppm_indepnormal.R
\name{graphppm_indepnormal}
\alias{graphppm_indepnormal}
\title{GraphPPM with indpendent normal node attributes}
\usage{
graphppm_indepnormal(
  graph0,
  Y,
  logcohesion,
  cohesion_param = NULL,
  mu0 = rep(0, ncol(Y)),
  kappa0 = rep(1, ncol(Y)),
  a0 = rep(1, ncol(Y)),
  b0 = rep(1, ncol(Y)),
  nsave = 1000,
  nburn = 1000,
  nthin = 1,
  z_init = NULL
)
}
\arguments{
\item{graph0}{\emph{igraph} igraph object, base graph}

\item{Y}{\emph{matrix<int> (n,d)} each row corresponds to d independent normal node attributes.}

\item{logcohesion}{\emph{function} logcohesion function}

\item{cohesion_param}{\emph{list} cohesion function parameters}

\item{mu0}{\emph{vector<num> (d)}, normal-invgamma location parameter for each dimension}

\item{kappa0}{\emph{vector<num> (d)}, normal-invgamma parameter for each dimension, where 1/sqrt(kappa0) is a scale parameter}

\item{a0}{\emph{vector<num> (d)}, invgamma shape parameter for each dimension}

\item{b0}{\emph{vector<num> (d)}, invgamma scale parameter for each dimension}

\item{nsave}{\emph{int} number of samples}

\item{nburn}{\emph{int} number of burn-in iterations}

\item{nthin}{\emph{int} number of thin-in rate. Total iterations: nburn + nsave*nthin}

\item{z_init}{\emph{vector<int> (n)}, initial partition}
}
\description{
Node attribute is modeled with independent normal, with normal-inversegamma prior.
i.e. Y_{il} ~ N(mu\link{l}, 1/lambda\link{l}), independently across i and l
and prior (see Murphy notes section 3)
mu\link{l} | lambda\link{l} ~ N(mu0\link{l}, 1/(kappa0\link{l}lambda\link{l})),
lambda\link{l} ~ gamma(a0\link{l}, b0\link{l})
}
