% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphppm_multinomial.R
\name{graphppm_multinomial}
\alias{graphppm_multinomial}
\title{GraphPPM categorical response analysis using MCMC}
\usage{
graphppm_multinomial(
  graph0,
  Y,
  logcohesion,
  cohesion_param = NULL,
  nclass = ncol(Y),
  a0 = 1,
  nsave = 1000,
  nburn = 1000,
  nthin = 1,
  z_init = NULL
)
}
\arguments{
\item{graph0}{\emph{igraph} igraph object, base graph}

\item{Y}{\emph{matrix<int> (n,d)} each row corresponds to d independent categorical node attributes. (NOTE:  this is not d classes)}

\item{logcohesion}{\emph{function} logcohesion function}

\item{cohesion_param}{\emph{list} cohesion function parameters}

\item{nclass}{\emph{vector<int> (d)}, number of classes for each categorical attribute}

\item{a0}{\emph{vector<num> (d)}, symmetric Dirichlet parameter for each categorical attribute}

\item{nsave}{\emph{int} number of samples}

\item{nburn}{\emph{int} number of burn-in iterations}

\item{nthin}{\emph{int} number of thin-in rate. Total iterations: nburn + nsave*nthin}

\item{z_init}{\emph{vector<int> (n)}, initial partition}
}
\description{
Node attribute is modeled with independent categorical distribution, with symmetric Dirichlet prior.
i.e. Y_{il} ~ Categorical(theta_1, ..., theta_{N_l}), independently across i and l, where N_l = nclass\link{l}
and prior (theta_1,...theta_{N_l}) ~ Dirichlet(a0\link{l}, ..., a0\link{l})
}
