% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphppm_prior.R
\name{graphppm_prior}
\alias{graphppm_prior}
\title{GraphPPM prior analysis using MCMC}
\usage{
graphppm_prior(
  graph0,
  logcohesion,
  cohesion_param = NULL,
  nsave = 1000,
  nburn = 1000,
  nthin = 1,
  z_init = NULL
)
}
\arguments{
\item{graph0}{\emph{igraph} igraph object, base graph}

\item{logcohesion}{\emph{function} logcohesion function}

\item{cohesion_param}{\emph{list} cohesion function parameters}

\item{nsave}{\emph{int} number of samples}

\item{nburn}{\emph{int} number of burn-in iterations}

\item{nthin}{\emph{int} number of thin-in rate. Total iterations: nburn + nsave*nthin}

\item{z_init}{\emph{vector<int> (n)}, initial partition}
}
\description{
GraphPPM prior analysis using MCMC
}
