% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphppm_prior_impsamp.R
\name{graphppm_prior_impsamp}
\alias{graphppm_prior_impsamp}
\title{GraphPPM prior analysis with importance sampler}
\usage{
graphppm_prior_impsamp(
  graph0,
  logcohesion,
  cohesion_param = NULL,
  nsave = 2000,
  kvals = NULL,
  kprobs = NULL,
  nCores = 1
)
}
\arguments{
\item{graph0}{\emph{igraph} igraph object, base graph}

\item{logcohesion}{\emph{function} logcohesion function}

\item{cohesion_param}{\emph{list} cohesion function parameters}

\item{nsave}{\emph{int} number of samples}

\item{kvals}{\emph{vector<int>} range of k}

\item{kprobs}{*vector<num>*unnormalized probability of q(k)}

\item{nCores}{\emph{int} default 1, set 2 or more lead to parallel run}
}
\description{
GraphPPM prior analysis with importance sampler
}
