% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logcohesions.R
\name{logcohesion_CPM}
\alias{logcohesion_CPM}
\title{Cohesion based on constant Potts model (loss-based)}
\usage{
logcohesion_CPM(A, deg, cohesion_param, minus_logtau = T)
}
\arguments{
\item{A}{}

\item{deg}{}

\item{cohesion_param}{}

\item{minus_logtau}{}
}
\description{
log(c) where c = alpha * c0^psi, and
c0(G_j, \partial G_j) = resolution * |V_j|^2 - (2*|E_j|)
default resolution = quantile(degree(g))\link{2} / (degree(g) - 1) ? (See ?igraph::cluster_leiden)
require argument m, the number of edges in the original graph
}
