% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logcohesions.R
\name{logcohesion_gamma}
\alias{logcohesion_gamma}
\title{misc cohesion}
\usage{
logcohesion_gamma(A, deg = NULL, cohesion_param, minus_logtau = T)
}
\arguments{
\item{A}{}

\item{deg}{}

\item{cohesion_param}{}

\item{minus_logtau}{}
}
\description{
misc cohesion
}
