% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logcohesions.R
\name{logcohesion_modularity}
\alias{logcohesion_modularity}
\title{Modularity cohesion function (loss-based)}
\usage{
logcohesion_modularity(A, deg, cohesion_param, minus_logtau = T)
}
\arguments{
\item{A}{\emph{matrix<int> (nj,nj)}, adjacency matrix, can be a sparse format.}

\item{deg}{\emph{vec<int&gt}, vector of node degrees with length nj}

\item{cohesion_param}{\emph{list}, must include "psi", "resolution", and "m" parameters.}

\item{minus_logtau}{\emph{logical} if true, return log c(Gj) - tau(Gj)}
}
\description{
Return c(G_j, \partial G_j) = exp(-psi\emph{L)
where L = resolution * (2|E_j| + partial G_j)^2/(2m)^2 - (2}|E_j|)/(2m)
}
