% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logcohesions.R
\name{logcohesion_vetau}
\alias{logcohesion_vetau}
\title{spanning tree cohesion (tree-based)}
\usage{
logcohesion_vetau(A, deg = NULL, cohesion_param, minus_logtau = T)
}
\arguments{
\item{A}{}

\item{deg}{}

\item{cohesion_param}{}

\item{minus_logtau}{}
}
\description{
spanning tree cohesion (tree-based)
}
