% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotient_graph.R
\name{make_quotient_graph}
\alias{make_quotient_graph}
\title{Make the quotient graph modulo a partition}
\usage{
make_quotient_graph(g0, z, return.igraph = F)
}
\arguments{
\item{z}{\emph{vector<int> (n)}, cluster membership}

\item{return.igraph}{\emph{logical}, return igraph, otherwise edgelist. Note: igraph output has eid from the original graph}

\item{E0}{\emph{matrix<int> (m,2)}, edge list of the graph}
}
\value{
\emph{edge list (with 2 columns) or igraph object}, quotient graph modulo a partition. Edge list output contains rownames as an edge id.
}
\description{
Make G/z, a quotient graph of G modulo a partition z.
The nodes of G/z corresponds to clusters (blocks) of a partition, and edges corresponds to edges between clusters.
Resulting quotient graph is often a multigraph, i.e. can have more than one edges between two nodes.
}
\examples{
g0 = igraph::make_graph("Zachary")
z = c(1,1,1,1,1,1,1,1,2,2,1,1,1,1,2,2,1,1,2,1,2,1,2,2,2,2,2,2,2,2,2,2,2,2)
make_quotient_graph(g0, z)
}
