% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsptrees.R
\name{nsptrees}
\alias{nsptrees}
\alias{nsptrees_igraph}
\title{The number of spanning trees of a graph}
\usage{
nsptrees(A, log = T)

nsptrees_igraph(g, log = T, sparse = F)
}
\arguments{
\item{A}{\emph{matrix<int> (n,n)}, Adjacency matrix, can be a sparse format.}

\item{log}{\emph{logical}, If TRUE (default), returns log value.}

\item{g}{\emph{igraph object}, Graph}

\item{sparse}{\emph{logical}, If TRUE (default), uses sparse matrix calculation.}
}
\value{
\emph{double}, The (log of) number of spanning trees
}
\description{
Calculate the number of spanning trees of a graph, using adjacency matrix or igraph object
}
\section{Functions}{
\itemize{
\item \code{nsptrees_igraph()}: Input is replaced with an igraph object g.

}}
\examples{
# Example 1
g = igraph::make_graph("Diamond")
A = igraph::as_adj(g, sparse = F)
nsptrees(A, log = F)
nsptrees_igraph(g, log = F)

# Example 2
g = igraph::make_graph("Zachary")
A = igraph::as_adj(g, sparse = T)
nsptrees(A, log = T)
nsptrees_igraph(g, log = T, sparse = T)
}
