% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_sort.R
\name{partition_sort}
\alias{partition_sort}
\title{Sort a partition by first-appearance order}
\usage{
partition_sort(z)
}
\arguments{
\item{z}{\emph{vector<int> (n)}, unsorted partition, represented as cluster membership vector}
}
\value{
sorted partition (cluster membership vector)
}
\description{
Sort a partition (cluster membership vector) by the first-appearance order
That is, let t_j:=min\{i: z_i =j\}, then j<k implies t_j<t_k.
}
\examples{

# partition_sort(c(2,1,5,1,2))
# should be 1 2 3 2 1

}
