% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_sort.R
\name{partition_vaild}
\alias{partition_vaild}
\title{Check partition validity}
\usage{
partition_vaild(z)
}
\arguments{
\item{z}{\emph{vector<int> (n)}, partition}
}
\value{
\emph{logical}, FALSE if invalid
}
\description{
If partition with k clusters is represented as cluster labels, it must have labels 1,2,...k.
This does not require partition to be sorted in first-appearance order but must include labels 1 through k without missing labels between.
}
\examples{
partition_vaild(c(2,1,5,1,2)) # FALSE, since label 3 and 4 are missing
partition_vaild(c(2,1,3,1,2)) # TRUE
partition_vaild(partition_sort(c(2,1,3,1,2))) # always TRUE

}
