% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runifsptree.R
\name{runifsptree}
\alias{runifsptree}
\title{Draw a uniform spanning tree using Wilson's algorithm}
\usage{
runifsptree(g0, root = NULL, ordering = NULL, maxiter = NULL)
}
\arguments{
\item{g0}{\emph{igraph object}, Graph.}

\item{root}{\emph{int}, (optional) Root node. Recommended as a node with max degree}

\item{ordering}{\emph{vector<int> (n)}, (optional) Ordering of nodes. Recommended as a breadth-first search ordering at the root.}
}
\value{
\emph{matrix<int> (n-1,2)}, Edge list of the sampled tree.
}
\description{
Equivalent algorithm available in igraph package igraph::sample_spanning_tree(g) but this one should be much faster.
}
\examples{
g = igraph::make_graph("Zachary")
root = which.max(igraph::degree(g))
ordering = as.integer(igraph::bfs(g, root)$order)
ust = runifsptree(g, root, ordering)


}
