% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runifsptree.R
\name{runifsptree_given_z}
\alias{runifsptree_given_z}
\title{Draw a uniform spanning tree conditional on a partition using Wilson's algorithm}
\usage{
runifsptree_given_z(g0, z, return.igraph = T)
}
\arguments{
\item{g0}{\emph{igraph object}, Graph.}

\item{z}{\emph{vector<int> (n)}, partition}

\item{return.igraph}{\emph{logical}, return igraph, otherwise edgelist with THREE columns where last column corresponds to crossing status}
}
\value{
\emph{igraph or matrix<int> (n-1,3)},
}
\description{
Draw a uniform spanning tree conditional on a partition using Wilson's algorithm
}
\examples{
g0 = igraph::make_graph("Zachary")
z = c(1,1,1,1,1,1,1,1,2,2,1,1,1,1,2,2,1,1,2,1,2,1,2,2,2,2,2,2,2,2,2,2,2,2)
runifsptree_given_z(g0, z)

}
