% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphppm_prior.R
\name{graphppm_prior}
\alias{graphppm_prior}
\title{GraphPPM prior analysis using MCMC}
\usage{
graphppm_prior(
  graph0,
  logcohesion,
  cohesion_param = NULL,
  nsave = 1000,
  nburn = 1000,
  nthin = 1,
  z_init = NULL,
  ...
)
}
\arguments{
\item{graph0}{igraph object, base graph}

\item{logcohesion}{logcohesion function}

\item{cohesion_param}{cohesion function parameters}

\item{nsave}{number of samples}

\item{nburn}{}

\item{nthin}{}

\item{z_init}{}

\item{...}{}
}
\description{
GraphPPM prior analysis using MCMC
}
