% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphppm_prior_impsamp.R
\name{graphppm_prior_impsamp}
\alias{graphppm_prior_impsamp}
\title{GraphPPM prior analysis with importance sampler}
\usage{
graphppm_prior_impsamp(
  graph0,
  logcohesion,
  cohesion_param = NULL,
  nsave = 2000,
  kvals = NULL,
  kprobs = NULL,
  nCores = 1,
  ...
)
}
\arguments{
\item{graph0}{igraph object, base graph}

\item{logcohesion}{logcohesion function}

\item{cohesion_param}{cohesion function parameters}

\item{nsave}{number of samples}

\item{kvals}{range of k}

\item{kprobs}{unnormalized probability of q(k)}

\item{nCores}{default 1, set 2 or more lead to parallel run}

\item{...}{}
}
\description{
GraphPPM prior analysis with importance sampler
}
